Update Create-New-AD-User-O365-Synced.ps1
Changed $EmailUsername to $EmailConvention. It reads the values from the Params sheet and tests in the user creation loop
Previously, Powershell was using the literal string value instead of reading the variables.
Created If and ElseIf conditions to test for most common username formats, falling back to first initial + last name if nothing else is met.
Tested and it is working properly.