Fixed MSOL User variable and $Mailbox.isShared issue
Changed parameters to correct Firstname/Lastname for MSOL user query so it would bring up results.
Added new mailbox query so the "isshared" attribute would update, as it remains the same in the variable otherwise.
Tested and it now brings up the MSOL user, correctly detects that mailbox is shared, and removed the O365 license.